refactor: rename form api ref to form ref and pull out code into imperative handle hook
