+++ packages/widget/src/AppProvider.tsx
@@ -17,7 +17,7 @@ import type { WidgetConfigProps } from './types/widget.js';
 export const AppProvider: React.FC<PropsWithChildren<WidgetConfigProps>> = ({
   children,
   config,
-  formApiRef,
+  formRef,
 }) => {
   return (
     <QueryClientProvider client={queryClient}>
@@ -25,7 +25,7 @@ export const AppProvider: React.FC<PropsWithChildren<WidgetConfigProps>> = ({
         <WalletProvider>
           <ThemeProvider>
             <I18nProvider>
-              <StoreProvider config={config} formApiRef={formApiRef}>
+              <StoreProvider config={config} formRef={formRef}>
                 <AppRouter>{children}</AppRouter>
               </StoreProvider>
             </I18nProvider>
